requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
aiohttp
jinja2